requirements.txt
streamlit==1.50.0
tensorflow==2.20.0
numpy
opencv-python-headless
Pillow